requirements.txt
discord.py==1.7.3
aiohttp==3.8.1
websockets==10.4
colorama==0.4.6
